pyproject.toml
[build-system]
requires = ["hatchling>=1.27"]
build-backend = "hatchling.build"

[project]
name = "paper-wiki"
version = "0.4.0"
description = "Personal research wiki builder for Claude Code — pipeline-driven paper ingestion with plugin architecture."
readme = "README.md"
license = "GPL-3.0-only"
requires-python = ">=3.11"
authors = [{ name = "kaneyxx" }]
keywords = [
  "claude-code",
  "claude-code-plugin",
  "research-wiki",
  "personal-knowledge-base",
  "academic-tools",
  "paper-management",
  "obsidian",
  "arxiv",
]
classifiers = [
  "Development Status :: 3 - Alpha",
  "Environment :: Plugins",
  "Intended Audience :: Science/Research",
  "License :: OSI Approved :: GNU General Public License v3 (GPLv3)",
  "Operating System :: MacOS",
  "Operating System :: POSIX :: Linux",
  "Programming Language :: Python :: 3.11",
  "Programming Language :: Python :: 3.12",
  "Topic :: Scientific/Engineering",
  "Topic :: Documentation",
  "Framework :: AsyncIO",
]
dependencies = [
  "httpx>=0.27",
  "loguru>=0.7",
  "pymupdf>=1.24",
  "pydantic>=2.9",
  "pydantic-settings>=2.5",
  "pyyaml>=6.0",
  "typer>=0.12",
  "aiofiles>=24.1",
]

[project.optional-dependencies]
dev = [
  "ruff>=0.7",
  "mypy>=1.13",
  "pytest>=8.3",
  "pytest-asyncio>=0.24",
  "pytest-cov>=6.0",
  "vcrpy>=6.0",
  "syrupy>=4.7",
  "hypothesis>=6.115",
  "types-PyYAML>=6.0",
  "types-aiofiles>=24.1",
]

[project.scripts]
paperwiki = "paperwiki.cli:main"

[project.urls]
Homepage = "https://github.com/kaneyxx/paper-wiki"
Repository = "https://github.com/kaneyxx/paper-wiki"
Issues = "https://github.com/kaneyxx/paper-wiki/issues"

[tool.hatch.build.targets.wheel]
packages = ["src/paperwiki"]

[tool.ruff]
line-length = 100
target-version = "py311"
src = ["src", "tests"]

[tool.ruff.lint]
select = [
  "E",   # pycodestyle errors
  "F",   # pyflakes
  "I",   # isort
  "B",   # flake8-bugbear
  "UP",  # pyupgrade
  "N",   # pep8-naming
  "S",   # flake8-bandit (security)
  "RUF", # ruff-specific
  "PT",  # flake8-pytest-style
  "SIM", # flake8-simplify
  "TCH", # flake8-type-checking
  "PERF",# perflint
  "DTZ", # flake8-datetimez (catch naive datetimes)
]
ignore = [
  # Pydantic models need their type annotations at runtime, so moving
  # datetime imports into TYPE_CHECKING blocks would break validation.
  "TC003",
]

[tool.ruff.lint.per-file-ignores]
"tests/**" = ["S101", "S105", "S106", "S603", "S607"]

[tool.ruff.lint.isort]
known-first-party = ["paperwiki"]

[tool.mypy]
strict = true
python_version = "3.11"
files = ["src/paperwiki"]
plugins = ["pydantic.mypy"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
addopts = "--strict-markers"
markers = [
  "timeout: mark a test with a wall-clock budget (seconds)",
]

[tool.coverage.run]
source = ["paperwiki"]
branch = true

[tool.coverage.report]
exclude_lines = [
  "pragma: no cover",
  "if TYPE_CHECKING:",
  "raise NotImplementedError",
  "@overload",
]
